Working on MirrorPuzzle prefab...
Reordering not working
Hence Mirror uses child index broke when scene gets reodered?

+++ ChronoGame/Scripts/ElevatorMove.hpp
@@ -55,7 +55,7 @@ public:
 		// Called every frame while the script is enabled
 		if (isMoving)
 		{
-			moveTimer -= deltaTime;
+			moveTimer -= (float)deltaTime;
 			if (moveTimer <= 0.0f)
 			{
 				isMoving = false;
